⚡️ :: (#21) move cursor code to Dice.tsx from DiceRender.tsx
DiceRender 이름에 맞지 않는 커서 기능을 Dice.tsx로 가져왔다
