+++ functions/src/components/Test/Test.tsx
@@ -27,7 +27,6 @@ const Test = () => {
         <ambientLight intensity={0.2} />
         <Physics iterations={15} gravity={[0, -30, 0]} allowSleep={false}>
           <Plane color={color.lightBlue} position={[0, -5, 0]} rotation={[-Math.PI / 2, 0, 0]} />
-          <DiceRender />
         </Physics>
       </Canvas>
     </S.Container>
